refactor(concentrator): split session-store.ts into domain modules
session-store.ts: 3495 -> 3108 lines. Extracted into
src/concentrator/session-store/:
- parsers.ts (44) - pure helpers (context-mode, clipboard-mime, readable-text)
- terminal-registry.ts (62) - createTerminalRegistry() factory for viewer tracking
- channel-registry.ts (420) - createChannelRegistry(deps) for pub/sub with
  bidirectional indices + diagnostics

All 36 behavioral tests still pass. No public API changes.

Transcript cache + addEvent + persistence NOT extracted -- they would
need context objects with 8+ shared state refs plus callbacks, negating
readability wins. Left for follow-up pass with more careful design.

Also fixed test TaskInfo shape (added required updatedAt).
