+++ src/concentrator/session-store.ts
@@ -9,12 +9,10 @@ import { join } from 'node:path'
 import type { ServerWebSocket } from 'bun'
 import { resolveContextWindow } from '../shared/context-window'
 import type {
-  ChannelStats,
   HookEvent,
   LaunchConfig,
   Session,
   SessionSummary,
-  SubscriberDiag,
   SubscriptionChannel,
   SubscriptionsDiag,
   TaskInfo,
@@ -32,49 +30,12 @@ import type { UserGrant } from './permissions'
 import { resolvePermissionFlags, resolvePermissions } from './permissions'
 import { getProjectSettings } from './project-settings'
 import { appendSharedFile } from './routes'
+import { createChannelRegistry } from './session-store/channel-registry'
+import { detectClipboardMime, detectContextModeFromStdout, isReadableText } from './session-store/parsers'
+import { createTerminalRegistry } from './session-store/terminal-registry'
 import { listShares } from './shares'
 
 export type { SessionSummary }
-
-/** Parse /model or /context command stdout to detect 1M vs standard context mode.
- * Returns undefined if the entry isn't a relevant local-command-stdout. */
-function detectContextModeFromStdout(content: string): '1m' | 'standard' | undefined {
-  // Strip ANSI escape codes for cleaner matching
-  const clean = content.replace(/\u001b\[[0-9;]*m/g, '')
-  // /model confirmation: "Set model to <label>" or "Kept model as <label>"
-  const modelMatch = clean.match(/(?:Set model to|Kept model as)\s+(.+?)(?:\s+·|\n|$)/i)
-  if (modelMatch) {
-    return /\(1M context\)/i.test(modelMatch[1]) ? '1m' : 'standard'
-  }
-  // /context output: header "Context Usage" + full model id including variant suffix
-  if (/Context Usage/i.test(clean)) {
-    return /\[1m\]/i.test(clean) ? '1m' : 'standard'
-  }
-  return undefined
-}
-
-/** Detect image MIME type from base64 prefix (same logic as osc52-parser.ts) */
-function detectClipboardMime(base64: string): string | null {
-  if (base64.startsWith('iVBORw0K')) return 'image/png'
-  if (base64.startsWith('/9j/')) return 'image/jpeg'
-  if (base64.startsWith('R0lGOD')) return 'image/gif'
-  if (base64.startsWith('UklGR')) return 'image/webp'
-  return null
-}
-
-/** Check if decoded text is mostly printable (not garbled binary) */
-function isReadableText(text: string): boolean {
-  if (text.length === 0) return false
-  let printable = 0
-  for (let i = 0; i < text.length; i++) {
-    const code = text.charCodeAt(i)
-    // Printable ASCII, common Unicode, newlines, tabs
-    if ((code >= 0x20 && code < 0x7f) || code === 0x0a || code === 0x0d || code === 0x09 || code >= 0xa0) {
-      printable++
-    }
-  }
-  return printable / text.length > 0.8
-}
 
 // Dashboard broadcast message (concentrator -> browser)
 export interface DashboardMessage {
@@ -310,35 +271,10 @@ export function createSessionStore(options: SessionStoreOptions = {}): SessionSt
   // sessionId -> (wrapperId -> socket): multiple rclaude instances can share a Claude session
   const sessionSockets = new Map<string, Map<string, ServerWebSocket<unknown>>>()
   // Terminal viewers keyed by wrapperId (each PTY is on a specific wrapper)
-  const terminalViewers = new Map<string, Set<ServerWebSocket<unknown>>>()
+  const terminalRegistry = createTerminalRegistry()
   const dashboardSubscribers = new Set<ServerWebSocket<unknown>>()
-  const v2Subscribers = new Set<ServerWebSocket<unknown>>()
   let agentSocket: ServerWebSocket<unknown> | undefined
   let agentInfo: { machineId?: string; hostname?: string } | undefined
-
-  // Channel subscription registry (v2 pub/sub)
-  // Forward index: channel key -> set of subscriber sockets
-  const channelSubscribers = new Map<string, Set<ServerWebSocket<unknown>>>()
-  // Reverse index: socket -> subscriber info (channels, stats)
-  interface SubscriberEntry {
-    id: string
-    protocolVersion: number
-    connectedAt: number
-    channels: Map<
-      string,
-      {
-        channel: SubscriptionChannel
-        sessionId: string
-        agentId?: string
-        subscribedAt: number
-        messagesSent: number
-        bytesSent: number
-        lastMessageAt: number
-      }
-    >
-    totals: { messagesSent: number; bytesSent: number; messagesReceived: number; bytesReceived: number }
-  }
-  const subscriberRegistry = new Map<ServerWebSocket<unknown>, SubscriberEntry>()
   let subscriberIdCounter = 0
 
   // Sync protocol: epoch + monotonic sequence for message ordering and gap detection.
@@ -366,6 +302,25 @@ export function createSessionStore(options: SessionStoreOptions = {}): SessionSt
   function syncStamp(message: unknown): string {
     return JSON.stringify({ _epoch: SYNC_EPOCH, _seq: syncSeq, ...(message as Record<string, unknown>) })
   }
+
+  // Channel pub/sub registry -- created here so it can close over syncStamp + recordTraffic
+  // which are defined in this factory. dashboardSubscribers is a shared mutable ref.
+  const channelRegistry = createChannelRegistry({
+    dashboardSubscribers,
+    syncStamp,
+    recordTraffic,
+  })
+  const {
+    subscribeChannel,
+    unsubscribeChannel,
+    unsubscribeAllChannels,
+    getChannelSubscribers,
+    broadcastToChannel,
+    isV2Subscriber,
+    getSubscriptionsDiag,
+    migrateChannels,
+    clearSubagentChannels,
+  } = channelRegistry
 
   function sendSyncResponse(ws: ServerWebSocket<unknown>, type: string, extra?: Record<string, unknown>): void {
     ws.send(JSON.stringify({ type, epoch: SYNC_EPOCH, seq: syncSeq, ...extra }))
@@ -421,10 +376,6 @@ export function createSessionStore(options: SessionStoreOptions = {}): SessionSt
         break
       }
     }
-  }
-
-  function channelKey(channel: SubscriptionChannel, sessionId: string, agentId?: string): string {
-    return agentId ? `${channel}:${sessionId}:${agentId}` : `${channel}:${sessionId}`
   }
 
   // Pending agent descriptions: PreToolUse(Agent) pushes, SubagentStart pops
@@ -611,7 +562,7 @@ export function createSessionStore(options: SessionStoreOptions = {}): SessionSt
         ws.send(json)
         recordTraffic('out', json.length)
       } catch (err) {
-        const subInfo = subscriberRegistry.get(ws)
+        const subInfo = channelRegistry.getSubscriberEntry(ws)
         console.error(
           `[broadcast] Send failed to ${subInfo?.id || 'unknown'}: ${err instanceof Error ? err.message : err}`,
         )
@@ -633,7 +584,7 @@ export function createSessionStore(options: SessionStoreOptions = {}): SessionSt
         ws.send(json)
         recordTraffic('out', json.length)
       } catch (err) {
-        const subInfo = subscriberRegistry.get(ws)
+        const subInfo = channelRegistry.getSubscriberEntry(ws)
         console.error(
           `[broadcast] Send failed to ${subInfo?.id || 'unknown'}: ${err instanceof Error ? err.message : err}`,
         )
@@ -1191,73 +1142,12 @@ export function createSessionStore(options: SessionStoreOptions = {}): SessionSt
     }
 
     // Migrate channel subscriptions from oldId to newId
-    const channelTypes: SubscriptionChannel[] = [
-      'session:events',
-      'session:transcript',
-      'session:tasks',
-      'session:bg_output',
-    ]
-    for (const channel of channelTypes) {
-      const oldKey = channelKey(channel, oldId)
-      const subs = channelSubscribers.get(oldKey)
-      if (!subs || subs.size === 0) continue
-
-      const newKey = channelKey(channel, newId)
-      let newSubs = channelSubscribers.get(newKey)
-      if (!newSubs) {
-        newSubs = new Set()
-        channelSubscribers.set(newKey, newSubs)
-      }
-
-      for (const ws of subs) {
-        newSubs.add(ws)
-        // Update reverse index
-        const entry = subscriberRegistry.get(ws)
-        if (entry) {
-          const oldStats = entry.channels.get(oldKey)
-          entry.channels.delete(oldKey)
-          entry.channels.set(newKey, {
-            channel,
-            sessionId: newId,
-            subscribedAt: oldStats?.subscribedAt || Date.now(),
-            messagesSent: oldStats?.messagesSent || 0,
-            bytesSent: oldStats?.bytesSent || 0,
-            lastMessageAt: oldStats?.lastMessageAt || 0,
-          })
-        }
-        // Notify dashboard of rollover
-        try {
-          ws.send(
-            JSON.stringify({
-              type: 'channel_ack',
-              channel,
-              sessionId: newId,
-              status: 'subscribed',
-              previousSessionId: oldId,
-            }),
-          )
-        } catch {
-          /* dead socket, will be cleaned up */
-        }
-      }
-      channelSubscribers.delete(oldKey)
-    }
+    migrateChannels(oldId, newId)
 
     // Project links are CWD-based, no migration needed on rekey.
 
     // Clear subagent transcript subscriptions (subagents are reset on rekey)
-    for (const key of channelSubscribers.keys()) {
-      if (key.startsWith(`session:subagent_transcript:${oldId}:`)) {
-        const subs = channelSubscribers.get(key)
-        if (subs) {
-          for (const ws of subs) {
-            const entry = subscriberRegistry.get(ws)
-            if (entry) entry.channels.delete(key)
-          }
-        }
-        channelSubscribers.delete(key)
-      }
-    }
+    clearSubagentChannels(oldId)
 
     // Broadcast update (not end+create) so dashboard stays on this session
     broadcastSessionScoped(
@@ -1929,56 +1819,21 @@ export function createSessionStore(options: SessionStoreOptions = {}): SessionSt
     return wrappers ? Array.from(wrappers.keys()) : []
   }
 
-  // Terminal viewer management (multiple viewers per session)
-  function addTerminalViewer(wrapperId: string, ws: ServerWebSocket<unknown>): void {
-    let viewers = terminalViewers.get(wrapperId)
-    if (!viewers) {
-      viewers = new Set()
-      terminalViewers.set(wrapperId, viewers)
-    }
-    viewers.add(ws)
-  }
-
-  const EMPTY_VIEWER_SET: Set<ServerWebSocket<unknown>> = new Set()
-  function getTerminalViewers(wrapperId: string): Set<ServerWebSocket<unknown>> {
-    return terminalViewers.get(wrapperId) || EMPTY_VIEWER_SET
-  }
-
-  function removeTerminalViewer(wrapperId: string, ws: ServerWebSocket<unknown>): void {
-    const viewers = terminalViewers.get(wrapperId)
-    if (viewers) {
-      viewers.delete(ws)
-      if (viewers.size === 0) terminalViewers.delete(wrapperId)
-    }
-  }
-
-  function removeTerminalViewerBySocket(ws: ServerWebSocket<unknown>): void {
-    for (const [id, viewers] of terminalViewers) {
-      viewers.delete(ws)
-      if (viewers.size === 0) terminalViewers.delete(id)
-    }
-  }
-
-  function hasTerminalViewers(wrapperId: string): boolean {
-    const viewers = terminalViewers.get(wrapperId)
-    return !!viewers && viewers.size > 0
-  }
+  // Terminal viewer management (multiple viewers per session) -- delegated to terminal-registry
+  const {
+    addTerminalViewer,
+    getTerminalViewers,
+    removeTerminalViewer,
+    removeTerminalViewerBySocket,
+    hasTerminalViewers,
+  } = terminalRegistry
 
   // Dashboard subscriber management
   function addSubscriber(ws: ServerWebSocket<unknown>, protocolVersion = 1): void {
     dashboardSubscribers.add(ws)
 
-    // Track v2 subscribers and create registry entry
-    if (protocolVersion >= 2) {
-      v2Subscribers.add(ws)
-    }
-    subscriberRegistry.set(ws, {
-      id: `ws-${++subscriberIdCounter}`,
-      protocolVersion,
-      connectedAt: Date.now(),
-      channels: new Map(),
-      totals: { messagesSent: 0, bytesSent: 0, messagesReceived: 0, bytesReceived: 0 },
-    })
+    // Track v2 subscribers and create registry entry (delegated to channel registry)
+    channelRegistry.registerSubscriber(ws, protocolVersion, () => ++subscriberIdCounter)
 
     sendSessionsList(ws)
 
@@ -2018,249 +1873,12 @@ export function createSessionStore(options: SessionStoreOptions = {}): SessionSt
   function removeSubscriber(ws: ServerWebSocket<unknown>): void {
     const wasShareViewer = !!(ws.data as { shareToken?: string }).shareToken
     dashboardSubscribers.delete(ws)
-    v2Subscribers.delete(ws)
-    unsubscribeAllChannels(ws)
-    subscriberRegistry.delete(ws)
+    // Unregister from channel registry (removes v2, unsubscribes all channels, deletes registry entry)
+    channelRegistry.unregisterSubscriber(ws)
 
     // If a share viewer disconnected, notify admins about updated viewer counts
     if (wasShareViewer) {
       broadcastSharesUpdate()
-    }
-  }
-
-  // Channel subscription management (v2 pub/sub)
-  function subscribeChannel(
-    ws: ServerWebSocket<unknown>,
-    channel: SubscriptionChannel,
-    sessionId: string,
-    agentId?: string,
-  ): void {
-    const key = channelKey(channel, sessionId, agentId)
-    let subs = channelSubscribers.get(key)
-    if (!subs) {
-      subs = new Set()
-      channelSubscribers.set(key, subs)
-    }
-    subs.add(ws)
-
-    // Track in reverse index
-    const entry = subscriberRegistry.get(ws)
-    if (entry) {
-      entry.channels.set(key, {
-        channel,
-        sessionId,
-        agentId,
-        subscribedAt: Date.now(),
-        messagesSent: 0,
-        bytesSent: 0,
-        lastMessageAt: 0,
-      })
-    }
-  }
-
-  function unsubscribeChannel(
-    ws: ServerWebSocket<unknown>,
-    channel: SubscriptionChannel,
-    sessionId: string,
-    agentId?: string,
-  ): void {
-    const key = channelKey(channel, sessionId, agentId)
-    const subs = channelSubscribers.get(key)
-    if (subs) {
-      subs.delete(ws)
-      if (subs.size === 0) channelSubscribers.delete(key)
-    }
-
-    const entry = subscriberRegistry.get(ws)
-    if (entry) entry.channels.delete(key)
-  }
-
-  function unsubscribeAllChannels(ws: ServerWebSocket<unknown>): void {
-    const entry = subscriberRegistry.get(ws)
-    if (!entry) return
-
-    for (const key of entry.channels.keys()) {
-      const subs = channelSubscribers.get(key)
-      if (subs) {
-        subs.delete(ws)
-        if (subs.size === 0) channelSubscribers.delete(key)
-      }
-    }
-    entry.channels.clear()
-  }
-
-  function getChannelSubscribers(
-    channel: SubscriptionChannel,
-    sessionId: string,
-    agentId?: string,
-  ): Set<ServerWebSocket<unknown>> {
-    const key = channelKey(channel, sessionId, agentId)
-    return channelSubscribers.get(key) || new Set()
-  }
-
-  function broadcastToChannel(
-    channel: SubscriptionChannel,
-    sessionId: string,
-    message: unknown,
-    agentId?: string,
-  ): void {
-    // Channel messages are per-session, not buffered for sync catchup.
-    // Clients re-subscribe on reconnect and get fresh initial data.
-    const json = syncStamp(message)
-    const bytes = json.length
-    const sent = new Set<ServerWebSocket<unknown>>()
-
-    // Pre-compute filtered JSON for share viewers with hideUserInput
-    // Only needed for transcript channels (user entries are in transcript_entries messages)
-    let filteredJson: string | null = null
-    let filteredBytes = 0
-    function getFilteredJson(): string {
-      if (filteredJson !== null) return filteredJson
-      const msg = message as { entries?: Array<{ type?: string }> }
-      if (msg.entries) {
-        const filtered = { ...msg, entries: msg.entries.filter(e => e.type !== 'user') }
-        // Skip sending if all entries were user messages
-        filteredJson = filtered.entries.length > 0 ? syncStamp(filtered) : ''
-        filteredBytes = filteredJson.length
-      } else {
-        filteredJson = json // no entries field, pass through
-        filteredBytes = bytes
-      }
-      return filteredJson
-    }
-
-    // Send to v2 channel subscribers
-    const key = channelKey(channel, sessionId, agentId)
-    const subs = channelSubscribers.get(key)
-    if (subs) {
-      for (const ws of subs) {
-        try {
-          // Filter transcript entries for share viewers with hideUserInput
-          const wsData = ws.data as { hideUserInput?: boolean }
-          if (channel === 'session:transcript' && wsData.hideUserInput) {
-            const fj = getFilteredJson()
-            if (!fj) {
-              sent.add(ws)
-              continue // all entries were user messages, skip
-            }
-            ws.send(fj)
-            sent.add(ws)
-            recordTraffic('out', filteredBytes)
-            const entry = subscriberRegistry.get(ws)
-            if (entry) {
-              entry.totals.messagesSent++
-              entry.totals.bytesSent += filteredBytes
-              const chStats = entry.channels.get(key)
-              if (chStats) {
-                chStats.messagesSent++
-                chStats.bytesSent += filteredBytes
-                chStats.lastMessageAt = Date.now()
-              }
-            }
-            continue
-          }
-          ws.send(json)
-          sent.add(ws)
-          recordTraffic('out', bytes)
-          // Track per-channel stats
-          const entry = subscriberRegistry.get(ws)
-          if (entry) {
-            entry.totals.messagesSent++
-            entry.totals.bytesSent += bytes
-            const chStats = entry.channels.get(key)
-            if (chStats) {
-              chStats.messagesSent++
-              chStats.bytesSent += bytes
-              chStats.lastMessageAt = Date.now()
-            }
-          }
-        } catch (err) {
-          const subInfo = subscriberRegistry.get(ws)
-          console.error(
-            `[broadcast] Send failed to ${subInfo?.id || 'unknown'}: ${err instanceof Error ? err.message : err}`,
-          )
-          subs.delete(ws)
-          if (subs.size === 0) channelSubscribers.delete(key)
-        }
-      }
-    }
-
-    // Also send to legacy (v1) subscribers that haven't received it
-    for (const ws of dashboardSubscribers) {
-      if (!sent.has(ws) && !v2Subscribers.has(ws)) {
-        try {
-          ws.send(json)
-          recordTraffic('out', bytes)
-          const entry = subscriberRegistry.get(ws)
-          if (entry) {
-            entry.totals.messagesSent++
-            entry.totals.bytesSent += bytes
-          }
-        } catch (err) {
-          const subInfo = subscriberRegistry.get(ws)
-          console.error(
-            `[broadcast] Send failed to ${subInfo?.id || 'unknown'}: ${err instanceof Error ? err.message : err}`,
-          )
-          dashboardSubscribers.delete(ws)
-        }
-      }
-    }
-  }
-
-  function isV2Subscriber(ws: ServerWebSocket<unknown>): boolean {
-    return v2Subscribers.has(ws)
-  }
-
-  function getSubscriptionsDiag(): SubscriptionsDiag {
-    const subscribers: SubscriberDiag[] = []
-    for (const [ws, entry] of subscriberRegistry) {
-      const channels: ChannelStats[] = []
-      for (const ch of entry.channels.values()) {
-        channels.push({
-          channel: ch.channel,
-          sessionId: ch.sessionId,
-          agentId: ch.agentId,
-          subscribedAt: ch.subscribedAt,
-          messagesSent: ch.messagesSent,
-          bytesSent: ch.bytesSent,
-          lastMessageAt: ch.lastMessageAt,
-        })
-      }
-      const wsData = ws.data as { userName?: string } | undefined
-      subscribers.push({
-        id: entry.id,
-        userName: wsData?.userName,
-        protocolVersion: entry.protocolVersion,
-        connectedAt: entry.connectedAt,
-        channels,
-        totals: { ...entry.totals },
-      })
-    }
-
-    // Channel counts summary
-    const channelCounts: Record<string, number> = {}
-    for (const [key, subs] of channelSubscribers) {
-      const channelName = key.split(':').slice(0, 2).join(':')
-      channelCounts[channelName] = (channelCounts[channelName] || 0) + subs.size
-    }
-
-    let totalBytesSent = 0
-    let totalMessagesSent = 0
-    for (const entry of subscriberRegistry.values()) {
-      totalBytesSent += entry.totals.bytesSent
-      totalMessagesSent += entry.totals.messagesSent
-    }
-
-    return {
-      subscribers,
-      summary: {
-        totalSubscribers: dashboardSubscribers.size,
-        legacySubscribers: dashboardSubscribers.size - v2Subscribers.size,
-        v2Subscribers: v2Subscribers.size,
-        channelCounts,
-        totalBytesSent,
-        totalMessagesSent,
-      },
     }
   }
